backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
strawberry-graphql>=0.96.0
pandas>=1.3.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1
openpyxl>=3.0.0
python-multipart>=0.0.5
alembic>=1.7.0